refactor: rename `interfaces/` dir to `types/`
